src/addon/requirements.txt
fake-bpy-module-4.0==20231118